requirements.txt
# Core FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.23.2

# Data Handling and Numerics
pandas==2.1.1
numpy==1.26.0

# Machine Learning - Scikit-learn and XGBoost
scikit-learn==1.3.2
xgboost==2.0.0

# Model Explainability
shap==0.43.0
lime==0.2.0.1

# Bias and Fairness
aif360==0.6.0
# Note: AIF360 can have complex dependencies. 
# It might require specific versions of TensorFlow or other packages.
# Ensure compatibility with your Python environment (e.g., Python 3.8, 3.9 recommended for AIF360)
# Consider TensorFlow (often a dependency or used with AIF360, but not directly imported in current code)
# tensorflow # Uncomment and specify version if needed by AIF360 usage or other components

# CLI Framework
typer[all]==0.9.0

# Utilities
python-dotenv==1.0.0 # For managing environment variables if used (not explicitly in current code but good practice)
joblib==1.3.2 # For saving/loading scikit-learn models

# For alternative data (example, adjust as needed)
# requests==2.31.0 # If fetching data from web APIs
# beautifulsoup4==4.12.2 # If parsing HTML for digital footprints

# Ensure all versions are compatible with each other.
# It's good practice to test in a clean virtual environment.